--- benches/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod alloy_helpers;
-pub mod ethers_helpers;
