+++ examples/advanced/Cargo.toml
@@ -25,20 +25,3 @@ eyre.workspace = true
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
-
-
-[[bin]]
-name = "alloy_profit"
-path = "examples/uniswap_u256/alloy_profit.rs"
-
-[[bin]]
-name = "alloy_simulation"
-path = "examples/uniswap_u256/alloy_simulation.rs"
-
-[[bin]]
-name = "ethers_profit"
-path = "examples/uniswap_u256/ethers_profit.rs"
-
-[lib]
-name = "uniswap_u256"
-path = "examples/uniswap_u256/lib.rs"
