--- examples/advanced/examples/uniswap_u256/lib.rs
@@ -1,4 +1,0 @@
-//! Alloy and ethers helpers
-
-/// Helpers
-pub mod helpers;
