+++ examples/advanced/examples/uniswap_u256_alloy_profit.rs
@@ -1,10 +1,10 @@
 //! Uniswap V2 Arbitrage Profit Calculation using alloy U256
 
 use alloy::primitives::utils::format_units;
-use eyre::Result;
-use uniswap_u256::helpers::alloy::{
+use examples_advanced::helpers::alloy::{
     get_amount_in, get_amount_out, get_sushi_pair, get_uniswap_pair,
 };
+use eyre::Result;
 
 fn main() -> Result<()> {
     let uniswap_pair = get_uniswap_pair();
