+++ examples/advanced/src/helpers/mod.rs
@@ -1,4 +1,7 @@
+//! Helper modules for advanced examples
+
 /// Alloy helpers
 pub mod alloy;
+
 /// Ethers helpers
 pub mod ethers;
