+++ examples/advanced/src/lib.rs
@@ -1,0 +1,3 @@
+//! Helper modules for advanced examples
+
+pub mod helpers;
